% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustPlus.R
\name{ClustPlus}
\alias{ClustPlus}
\title{A clustering function}
\usage{
ClustPlus(
  mat = distribution_test_mat(nrow_x = 50, n_random_distributions = 100, class_method =
    "discovery"),
  autoscale_mat = T,
  center_fun = colMeans,
  scale_fun = colSds,
  color_clust_Nr = 5,
  nboots = 100,
  AU_p_value = 0.95,
  out_PDF = T,
  out_PDF_name = "test"
)
}
\arguments{
\item{mat}{Defaults to a randomly generated matrix from the distribution_test_mat function.}

\item{autoscale_mat}{Defaults to True, allows to center and scale the input matrix.}

\item{center_fun}{Defaults to colMeans, allows to select any function that can be applied to columns as condition to center the matrix.}

\item{scale_fun}{Defaults to colSds, allows to select any function that can be applied to columns as condition to scale the matrix.}

\item{color_clust_Nr}{Defaults to 5, is the number of colored-partitions that the decision dendrograms will have.}

\item{nboots}{Defaults to 100, is the number of bootstrapps using the package pvclust.}

\item{AU_p_value}{Defaults to 0.95 or in other words the classic P value 0.05 treshold to define a significant cluster.}

\item{out_PDF}{Defaults to TRUE, if TRUE generates the respective bootstrapped-dendrogram plots.}

\item{out_PDF_name}{Defaults to test, name of the PDF file that will contain the plots.}
}
\description{
This function allows you to rationalize the clustering method selected.
}
\examples{
test <- ClustPlus()
}
\keyword{bootstrapping,}
\keyword{class}
\keyword{clustering,}
\keyword{discovery,}
\keyword{pvclust}
