% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_test_mat.R
\name{distribution_test_mat}
\alias{distribution_test_mat}
\title{A Random distributions Function}
\usage{
distribution_test_mat(nrow_x = 1000, n_random_distributions = 1000)
}
\arguments{
\item{nrow_x}{Defaults to 1000.}

\item{n_random_distributions}{Defaults to 1000.}
}
\description{
This function allows you to generate a matrix with n randomly-drawn specific distribution shapes that follow GLM supported regression families.
}
\examples{
test_mat <- distribution_test_mat()
}
\keyword{distributions}
