% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsUnivariateStats.R
\name{OmicsUnivariateStats}
\alias{OmicsUnivariateStats}
\title{A Random distributions Function}
\usage{
OmicsUnivariateStats(
  class_comparison_mat = abs(distribution_test_mat()),
  Factor1,
  Factor2 = NULL,
  Contrast = T
)
}
\arguments{
\item{class_comparison_mat}{Defaults to distribution_test_mat().}

\item{Factor2}{defaults to NULL.}

\item{Contrast}{defaults to TRUE.}
}
\description{
This function allows you to test response variables using a generalized linear model with one or two factors and multiple levels per factor, i.e., multiple regressors, customizing the family of regression for each test according to the response variable distribution.
}
\examples{
...
}
\keyword{Statistics}
\keyword{Univariate}
