% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlusPCA.R
\name{PlusPCA}
\alias{PlusPCA}
\title{A clustering function}
\usage{
PlusPCA(
  inMat = distribution_test_mat(),
  Km = NULL,
  returnValuesPlot = T,
  returnTreatmentValues = T,
  pamClust = T,
  Kpam
)
}
\arguments{
\item{inMat}{Defaults to a randomly generated matrix from the distribution_test_mat function.}

\item{Km}{must be a result object from the function KmeansPlus().}

\item{returnValuesPlot}{if TRUE, returns all relevant PCA stats to manually build the plots.}

\item{returnTreatmentValues}{if TRUE, returns treatment values, if FALSE, returns features values.}
}
\description{
This function allows you to rationalize the clustering method selected.
}
\examples{
test <- PlusPCA()
}
\keyword{PCA,}
\keyword{class}
\keyword{discovery,}
\keyword{prcomp}
