% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_distributions.R
\name{testing_distributions}
\alias{testing_distributions}
\title{A Random distributions Function}
\usage{
testing_distributions(Distribution_test_mat = distribution_test_mat())
}
\arguments{
\item{Distribution_test_mat}{Defaults to a matrix built with distribution_test_mat().}
}
\description{
This function allows you to test the distributions of response variables and output per response variable the nearest GLM family. The assumption is that the mean kurtosis and square of skewness represent the average of the distributions.
}
\examples{
test_mat_distributions <- testing_distributions()
}
\keyword{distributions}
\keyword{test}
