% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyCustomized.R
\name{TukeyCustomized}
\alias{TukeyCustomized}
\title{A Tukey HSD visual representation}
\usage{
TukeyCustomized(variable, factor, conf.level = 0.95, MainTitle = "")
}
\arguments{
\item{variable}{is a numeric vector outlining a response variable.}

\item{factor}{is a factor type of vector of the same lenght as variable.}

\item{conf.level}{Defaults to 0.95 CI.}

\item{MainTitle}{allows to insert a title for the returned plot.}
}
\description{
This function allows you to rationalize the clustering method selected.
}
\keyword{TukeyHSD}
\keyword{class}
\keyword{discovery,}
