% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variables2Shapes.R
\name{Variables2Shapes}
\alias{Variables2Shapes}
\title{A Random distributions Function}
\usage{
Variables2Shapes(
  Distribution_test_mat = RandoDiStats::distribution_test_mat(class_method =
    "discovery")
)
}
\arguments{
\item{Distribution_test_mat}{Defaults to a matrix built with distribution_test_mat().}
}
\description{
This function allows you to infer the distribution shapes of response variables as a two column matrix with the kurtosis and square of skeweness.
}
\examples{
test_mat_distribution_shapes <- Variables2Shapes()
}
\keyword{distributions}
\keyword{shapes}
