% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsUnivariateStats.R
\name{OmicsUnivariateStats}
\alias{OmicsUnivariateStats}
\title{A Random distributions Function to perform univariate statistical inference}
\usage{
OmicsUnivariateStats(
  class_comparison_mat = abs(RandoDiStats::distribution_test_mat()),
  Factor1,
  Factor2 = NULL,
  Contrast = F,
  TukeyReturns = c("MeanComparisons", "Letters"),
  ReturnTukeyPlots = T,
  TukeyPDFName = "test",
  marginsTukey = c(6, 12, 3, 3),
  returnObject = c("OmicsTests", "class_comparison_mat")
)
}
\arguments{
\item{class_comparison_mat}{Defaults to distribution_test_mat().}

\item{Factor2}{defaults to NULL.}

\item{Contrast}{defaults to TRUE.}

\item{TukeyReturns}{Tukey HSD can return either "MeanComparisons" or "Letters".}

\item{ReturnTukeyPlots}{defaults to TRUE, returns a PDF with the plots.}

\item{TukeyPDFName}{defaults to "test".}

\item{marginsTukey}{margins of the Tukey HSD plots}

\item{returnObject}{returns either the results from the "OmicsTests" or the cropped tested "class_comparison_mat"}

\item{Factor1.}{Needs to be defined}
}
\description{
This function allows you to test response variables using a generalized linear model with one or two factors and multiple levels per factor, i.e., multiple regressors, customizing the family of regression for each test according to the response variable distribution.
}
\examples{
Factor1_eg <- as.factor(c(rep("RED", 200), rep("GREEN", 200), rep("BLACK", 200),rep("WHITE", 200), rep("YELLOW", 200)))
test_OUS <- OmicsUnivariateStats(Factor1 = Factor1_eg)
...
}
\keyword{Statistics}
\keyword{Univariate}
