% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_distributions.R
\name{plotting_distributions}
\alias{plotting_distributions}
\title{A Random distributions Function}
\usage{
plotting_distributions(
  test_mat = distribution_test_mat(),
  vector_colors = "grey",
  transparency = 90,
  variable_name = "test_mat",
  ylim_plot = c(max(Variables2Shapes(test_mat)[, 2]), min(Variables2Shapes(test_mat)[,
    2])),
  xlim_plot = c(min(Variables2Shapes(test_mat)[, 1]), max(Variables2Shapes(test_mat)[,
    1]))
)
}
\arguments{
\item{test_mat}{Defaults to a matrix built with distribution_test_mat().}

\item{vector_colors}{Defaults to grey.}

\item{transparency}{Defaults to 90\%.}

\item{variable_name}{Defaults to test_mat.}

\item{ylim_plot}{Defaults to minimum and maximum kurtosis of the provided response variables.}

\item{xlim_plot}{Defaults to minimum and maximum square of skewness of the provided response variables.}
}
\description{
This function allows you to plot distributions from a matrix of response variables.
}
\examples{
plotting_distributions()
}
\keyword{distribution}
\keyword{plot}
